refactor: move TTLCache out of the cogs package
TTLCache is not a cog or an extension: it defines no commands or listeners,
never touches the bot, and has no setup() entry point. Living in bridger/cogs/
implied it was one of the modules load_extension pulls in at startup.

Move it to bridger/cache.py alongside the other shared infrastructure so the
cogs package holds only the two real extension modules.
